inventory-system/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langextract==1.1.1
google-cloud-vision==3.5.0
pyzbar==0.1.9
Pillow==10.2.0
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25
